requirements.txt
greek-accentuation==0.8
PyYAML==3.10
